+++ kamehouse-shell/bin/kamehouse/mariadb-csv-kamehouse.sh
@@ -47,7 +47,7 @@ executeExport() {
   else
     OUT_FILE_BASE="C:/Users/"${USER}"/home-synced/mariadb/csv/"
   fi
-  mariadb -u nikolqs -p${MARIADB_PASS_NIKOLQS} -e"set @outFileBase = '${OUT_FILE_BASE}'; `cat ${PATH_SQL}/csv-kamehouse.sql`"
+  mariadb -u kamehouse -p${MARIADB_PASS_KAMEHOUSE} -e"set @outFileBase = '${OUT_FILE_BASE}'; `cat ${PATH_SQL}/csv-kamehouse.sql`"
   checkCommandStatus "$?"
   if ${IS_LINUX_HOST}; then
     log.info "Moving generated csv files from ${TMP_EXPORT_DIR} to ${PATH_CSV}"