+++ kamehouse-shell/bin/kamehouse/mariadb-restore-kamehouse.sh
@@ -15,8 +15,8 @@ MARIADB_DUMP_FILE=${HOME}/home-synced/mariadb/dump/dump-kamehouse.sql
 mainProcess() {
   if [ -f "${MARIADB_DUMP_FILE}" ]; then
     log.info "Restoring kamehouse database from ${MARIADB_DUMP_FILE}"
-    mariadb -u nikolqs -p${MARIADB_PASS_NIKOLQS} kameHouse < ${MARIADB_DUMP_FILE}
-    checkCommandStatus "$?"
+    mariadb -u kamehouse -p${MARIADB_PASS_KAMEHOUSE} kamehouse < ${MARIADB_DUMP_FILE}
+    checkCommandStatus "$?" "Error loading dump-kamehouse.sql"
     log.info "mariadb restore command completed successfully"
   else
     log.error "${MARIADB_DUMP_FILE} doesn't exist."