+++ kamehouse-shell/bin/kamehouse/sql/mariadb/csv-kamehouse.sql
@@ -1,4 +1,4 @@
-use kameHouse;
+use kamehouse;
 
 SET @tableName := 'booking_request';
 SET @outputFile := CONCAT(@outFileBase, @tableName, '.tmpcsv');